+++ SDQ/load.h
@@ -2,7 +2,7 @@
 #include <fstream>
 #include <string>
 using namespace std;
-void readData(string inFileName, double data[], int length){
+void readData(string inFileName, float data[], int length){
     ifstream inFile;
     inFile.open(inFileName.c_str());
     if (inFile.is_open()) {
@@ -12,7 +12,7 @@ void readData(string inFileName, double data[], int length){
     else{ cerr << "Can't find input file " << inFileName << endl;}
 }
 
-void readData3x3(string inFileName, double data[3][3]){
+void readData3x3(string inFileName, float data[3][3]){
     ifstream inFile;
     inFile.open(inFileName.c_str());
     if (inFile.is_open()){
@@ -24,8 +24,8 @@ void readData3x3(string inFileName, double data[3][3]){
     else{cerr << "Can't find input file " << inFileName << endl;}
 }
 
-void Tanspose(double arr[][3]){
-    double tmp;
+void Tanspose(float arr[][3]){
+    float tmp;
     for(int i=0; i<3; i++){
         for(int j=i; j<3; j++){
             tmp = arr[i][j];
@@ -35,7 +35,7 @@ void Tanspose(double arr[][3]){
     }
 }
 
-void LoadSenMap(string model, double Sen_Map[3][64]){
+void LoadSenMap(string model, float Sen_Map[3][64]){
     if(model != "NoModel"){
         readData("./SenMap/"+model+"_Y_KLT.txt", Sen_Map[0], 64);
         readData("./SenMap/"+model+"_Cb_KLT.txt", Sen_Map[1], 64);
@@ -50,7 +50,7 @@ void LoadSenMap(string model, double Sen_Map[3][64]){
     }
 }
 
-void LoadColorConvW(string model, double W_rgb2swx[3][3], double W_swx2rgb[3][3]){
+void LoadColorConvW(string model, float W_rgb2swx[3][3], float W_swx2rgb[3][3]){
     if(model != "NoModel"){
         readData3x3("./color_conv_W/"+model+"_W_OPT.txt", W_rgb2swx);
         readData3x3("./color_conv_W/"+model+"_W_OPT.txt", W_swx2rgb);
